Fix "short" Following Current speed

+++ GameServer/gameobjects/GameNPC.cs
@@ -1747,7 +1747,7 @@ namespace DOL.GS
             if (InCombat || Brain is BomberBrain)
                 WalkTo(newX, newY, (ushort)newZ, MaxSpeed);
             else if (MaxSpeed < GetDistance(new Point2D(newX, newY)))
-                WalkTo(newX, newY, (ushort)newZ, Math.Min(MaxSpeed, followLiving.CurrentSpeed + 50));
+                WalkTo(newX, newY, (ushort)newZ, (short)Math.Min(MaxSpeed, followLiving.CurrentSpeed + 50));
             else
                 WalkTo(newX, newY, (ushort)newZ, (short)GetDistance(new Point2D(newX, newY)));
             return ServerProperties.Properties.GAMENPC_FOLLOWCHECK_TIME;
